#48 Added logger to MgcpConnection.
Renamed textualId to hexadecimalId on MgcpConnection.

Protecting underlying connection from MgcpConnection.

Applied proper lifecycle to pooled MgcpConnections.

MgcpConnection.endpoint is now final.

Guaranteed MgcpConnection objects are released when deleting all connections from an endpoint.

Added traces to control connection pool of MgcpEndpoint.

+++ controls/mgcp/src/main/java/org/mobicents/media/server/mgcp/tx/cmd/AuditConnectionCmd.java
@@ -223,8 +223,8 @@ public class AuditConnectionCmd extends Action {
 
         private void auditConnectionParameters(MgcpConnection connection) {
             connectionParameters = new ConnectionParameters();
-            connectionParameters.packetsSent = connection.getPacketsTransmitted();
-            connectionParameters.packetsReceived = connection.getPacketsReceived();
+            connectionParameters.packetsSent = (int) connection.getPacketsTransmitted();
+            connectionParameters.packetsReceived = (int) connection.getPacketsReceived();
             connectionParameters.jitter = (int) connection.getConnection().getJitter();
             // TODO hrosa - get transmitted octets from MGCP connection
             // TODO hrosa - get received octets from MGCP connection
